[BANKCON-11757] Allow keeping footer above keyboard in `PaneLayoutView` (#3910)
## Summary

This tweaks the `PaneLayoutView` to support keeping the footer view
always above the keyboard view. This behavior is **disabled by default**
when using a PaneLayoutView, as to not alter any existing screens. This
PR enables this on the `LinkLoginViewController`, which is currently
gated behind an unreleased feature.

| Before | After |
|--------|--------|
| ![Simulator Screenshot - iPhone 15 - 2024-08-08 at 14 39
40](https://github.com/user-attachments/assets/d701cf6b-8c25-4255-b7ff-452a0f7989b5)
| ![Simulator Screenshot - iPhone 15 - 2024-08-08 at 14 38
48](https://github.com/user-attachments/assets/c50588a1-bc72-459e-8391-109dfbd5cc9e)
|

> [!NOTE]  
> We intentionally don't move the footer view when the sheet is
presented on iPad, since the entire sheet is shifted above the keyboard
already.


https://github.com/user-attachments/assets/a98c38c7-e193-4eb3-a0f5-565d27ab00e8

## Motivation

✨ Product quality ✨ 

## Testing

> [!IMPORTANT]  
> I've done a bunch of manual testing myself, but I encourage reviewers
to think of edge cases here I might've missed.


https://github.com/user-attachments/assets/31ae77ca-0961-41d9-9921-efb802f87393

## Changelog

N/a