Dynamically set `request_surface` in FC API requests (#3947)
## Summary

For all API requests to the `/consumers/...` path, we need to provide
the `request_surface` body parameter. This is either
`ios_instant_debits` if we're in the Native Instant Debits flow or
`ios_connections` otherwise. This PR now dynamically sets that value
according to the manifest's `isLinkWithStripe` field (set
[here](https://stripe.sourcegraphcloud.com/github.com/stripe/stripe-ios/-/blob/StripeFinancialConnections/StripeFinancialConnections/Source/Native/NativeFlowDataManager.swift?L140)).

## Motivation

Consistency and correctness ✨ 

## Testing

N/a

## Changelog

N/a