[StripeConnect] Add example app. (#3969)
## Summary
This adds a basic example app utilizing the StripeConnect SDK.

[MXMOBILE-2480](https://jira.corp.stripe.com/browse/MXMOBILE-2480)

## App Breakdown
- On app open retrieve app_info from the API. If loading takes longer
than 1 second a reassurance message appears. On success push to the main
components view. On Failure show error message and option to retry or go
to app settings.
- The components list shows the Payouts option, more will be available
in the future. Selecting the payouts list will push on a
PayoutsViewController.
- Tapping the top nav bar takes you to app settings where you can
configure the server url and which account is currently selected.

## API Description
`GET: /app_info` : This is requested on app start and will return the
public key as well as the available accounts. The display name is
optional, but helpful.
Example Response:
```
{
  "publishable_key": "pk_test_<key>",
  "available_merchants": [
    {
      "display_name": "Name to display in app ",
      "merchant_id": "acct_123456"
    },
  ]
}
```
`POST /account_session` The will be called in fetchSession on the SDK.
If the account header is provided then it will get the session for that
account otherwise it will get a session from the default account which
is currently acct_1N9FIXQ26HdRlxHg.
Example Response:
```
{
   "client_secret": "<secret>"
}
```

## Demo


https://github.com/user-attachments/assets/ccaef189-8161-4670-a719-d9184cd1763e