MOBILESDK-1817 [iOS] Add validation to deferred flow to check Intent payment method id matches (#3972)
## Summary
<!-- Simple summary of what was changed. -->
Added validation functions to PaymentSheetDeferredValidator to check
Intent payment method id matches the id of the STPPaymentMethod passed
into the ConfirmHandler. Called this validator in 4a Client-side
confirmation and 4b Server-side confirmation in PaymentSheet+DeferredAPI
for both SetupIntent and PaymentIntent. Added simple tests for both.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
Based on this issue https://github.com/stripe/stripe-ios/issues/3358,
there's a bug merchants can make in their integration that won't get
caught today.

## Testing
<!-- How was the code tested? Be as specific as possible. -->
Reproduced the issue (Attempt Amazon Pay, cancel, and switch to card)
and made unit tests for the validator method with the following checks:
1. Ids match (does not throw)
2. Ids do not match (throws)
3. Payment method on intent is nil (does not throw)

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->