Add Appearance APIs for primary button customization in disabled state + snapshot tests + playground update (#3991)
## Summary
Added a disabledBackgroundColor and disabledTextColor to the Appearance
API for the primary button in its disabled state. Both are optional.
disabledBackgroundColor defaults to backgroundColor, which defaults to
the primary color if not set. disabledTextColor defaults to textColor
with an alpha component of 0.6.

## Motivation
Users have requested customization of the color of the primary button in
its disabled state.
https://jira.corp.stripe.com/browse/MOBILESDK-2418

## Testing
Snapshot tests: default, primary color set with no backgroundColor or
disabledBackgroundColor, backgroundColor set with no
disabledBackgroundColor, primary color, backgroundColor, and
disabledBackgroundColor all set, and just disabledBackgoundColor and
disabledTextColor set

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->