Fix checkbox color (#4019)
## Summary
The checkbox stroke color was set to
`appearance.colors.componentBorder`. This is reasonable, but if the
merchant sets `componentBorder` to clear to remove the border (which is
our recommended approach), it makes the checkbox disappear.

This PR makes the checkbox stroke color fall back to the secondary text
color if `componentBorder` is clear. And for Link, it makes the checkbox
stroke color always be the Link border color.

It also makes the checkbox background color = background color - that
looks better IMO, especially in the fallback case described above.

## Testing
**Before**
![CleanShot 2024-09-13 at 17 59
08@2x](https://github.com/user-attachments/assets/03afb635-b48b-4b2e-add7-44fa64140250)
**After**
![CleanShot 2024-09-13 at 17 59
20@2x](https://github.com/user-attachments/assets/788f4240-964d-41c3-81b2-c1dfe4e15f41)

## Changelog
[Fixed] Fixed an issue where checkboxes were not visible when
`appearance.colors.componentBorder` was transparent.