Alert on ambiguous layouts (#4044)
## Summary
In an attempt to clean up our ambiguous layout issues, try:
* Asserting on ambiguous layouts inside our snapshot tests
* In PaymentSheet Example, detect and shake ambiguous layout items. This
is off for now — it's too noisy, as we have a few known issues with
laying out hidden UIViews within UIStackViews. We'll turn it on once we
fix those.

(Thanks to @yuki-stripe for working out the original ambiguousView
detection technique!)

## Motivation
https://jira.corp.stripe.com/browse/RUN_MOBILESDK-3334

## Testing
Existing tests + manually in PaymentSheet Example

## Changelog
No user-facing changes.