[BANKCON-14528] Update FC SDK to accept `link_mode` value from the payment sheet (#4050)
## Summary

As part of the
[Panther](https://docs.google.com/document/d/1ErJVA3lLvNspPe3A8uYP9feK8Yvfq-Sw5aatNIvlGxk/edit#heading=h.p9ofrc4zbe1z)
project, we'll need to call a different payment method endpoint at the
end of the Financial Connections flow based on whether or not we're in a
Panther flow. To know this, we need the `link_mode` value, which is
fetched in the elements session response. This passes that value into
the FC SDK for future use!

## Motivation

Building Panther!

## Testing

Verified the correct `linkMode` value is accessed in the
`FinancialConnectionsSheet`:

<img width="1184" alt="Screenshot 2024-09-24 at 9 55 27 AM"
src="https://github.com/user-attachments/assets/5d0831a5-dbfb-4b58-836d-8a235267abde">

## Changelog

N/a