Fix ui-tests (#4208)
## Summary
<!-- Simple summary of what was changed. -->
Some UI tests were failing while waiting for a button that didn't exist,
and it's because it was expecting the horizontal layout but found the
vertical layout instead. Also increased the timeout for one case.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

https://app.bitrise.io/build/bdbb441b-becc-4c63-80b8-c394f2e3d2f3?tab=tests

https://app.bitrise.io/build/40c55a29-1cd9-4a90-b653-6188c5edad52?tab=tests

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->