Fix embeddedSettingsView displaying stale values by passing a closure (#4224)
## Summary
This PR addresses an issue where embeddedSettingsView in
PaymentSheetTestPlayground occasionally shows outdated values because
state changes weren't propagating to the embedded SwiftUI view within
UIKit. To fix this, we modified the presentEmbedded method to accept a
closure that returns a view (@escaping () -> SettingsView) instead of a
static view. This change ensures that embeddedSettingsView is
re-evaluated whenever it's accessed, allowing it to reflect the latest
state updates correctly.

## Motivation
- Embedded

## Testing
-  Manual

## Changelog
N/A