Added a flag for default spm navigation (#4271)
## Summary
<!-- Simple summary of what was changed. -->
Added a flag defaultSPMNavigation in PaymentSheet and CustomerSheet. Off
by default.
When on, edit icon shows for all cards regardless of cbc eligibility.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-2663](https://jira.corp.stripe.com/browse/MOBILESDK-2663)
[MOBILESDK-2664](https://jira.corp.stripe.com/browse/MOBILESDK-2664)
## Testing
<!-- How was the code tested? Be as specific as possible. -->
Simple UI tests that check to make sure that both the cbc card and
non-cbc card have edit buttons
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->