Fix Link signup in PaymentSheet when using a Connect account (#4319)
## Summary
When using a Connect account, we'd incorrectly pass the Stripe-Account
header alongside a consumer publishable key. We should follow the
strategy from https://github.com/stripe/stripe-ios/pull/4310/files and
disable the Stripe-Account header for these requests.

## Motivation
Fix issues with signing up for Link in Connect-using apps.

## Testing
Added a new test.

## Changelog
N/A, issue is not visible to users.