Add `PromoBadgeView` in horizontal and vertical elements (#4289)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds a `PromoBadge` component and renders it in the
payment method tab and payment method row. It’s currently only visible
in screenshot tests, but will start showing up in the sheet with
https://github.com/stripe/stripe-ios/pull/4285 (behind a feature flag).

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[CONSUMERBANK-568](https://jira.corp.stripe.com/browse/CONSUMERBANK-568)

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->