Fix incorrect `PaymentOptionDisplayData` for Link card brand (#4328)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue where the `PaymentOptionDisplayData`
showed incorrect information for payments with Link card brand.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Updated the UI test to validate the displayed data.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

[Fixed] Fixed an issue where FlowController showed incorrect
`PaymentOptionDisplayData` for Link card brand.