Read the default_payment_method field from elements session and use that as the default PM (#4313)
## Summary
<!-- Simple summary of what was changed. -->
Added allowsSetAsDefaultPM flag. Only shown if we're using customer
session.
If allowsSetAsDefaultPM is on, then instead of reading from the locally
stored default, we read from ElementsSession.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-2799](https://jira.corp.stripe.com/browse/MOBILESDK-2799)
## Testing
<!-- How was the code tested? Be as specific as possible. -->
customer_session, allowsSetAsDefaultPM on
Hardcoded the value that was "read" from default_payment_method (set to
the 4242 Visa stripe id)
After changing the selected payment method and paying with it, upon
reload the PaymentSheet default is back to our hardcoded default.



https://github.com/user-attachments/assets/38601a0e-513b-40a8-a3d8-f30069723f97



## Changelog


<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->