Fix MLModel loader (#4353)
## Summary
This code changed behavior after
https://github.com/stripe/stripe-ios/pull/3869: It's now running some
work on the main thread which should run on a background thread (such as
call to compile the model), and the temporary files sometimes fall out
of scope before they can be compiled. We'll move these calls to the
loadPromiseCacheQueue. It also feels brittle for
`downloadFileTemporarily` to rely on the `downloadTask`'s file sticking
around, so we'll make it explicit by moving it to a temporary directory.

## Motivation
Fix an issue with using StripeIdentity in SDKs after
https://github.com/stripe/stripe-ios/pull/3869 landed.

## Testing
In Identity Example.

## Changelog
Will add