Integrate App Attestation backend (#4369)
## Summary
Integrates App Attest using the backend APIs. This is less scary than it
looks, most of the changes are in the tests or just plumbing the
`use_attestation_endpoints` value down into Link.

### StripeAttest
* We're now observing the `initial_challenge_required` boolean as
another signal of whether to attest. If we're asked to attest an already
attested key (or not attest an unattested one), we'll log an error and
reset the key.
* We'll also reset the key if any asserted requests return an error. The
new `receivedAssertionError` function can be used to react to that
error.
* Added a new `prepareAttestation` function, which checks if we're able
to perform attestation and attempts to attest a key if we don't have one
already. This can be used to switch between the web and native flows,
depending on the likelihood that future assertions will succeed.

### PaymentSheet
* Removed the useNativeLink flag, as this behavior is now configured by
elements/sessions.
* We now pass the `mobile_app_id` to elements/sessions, enabling it to
decide whether to enable `link_mobile_use_attestation_endpoints`.
* Updated STPAPIClient+Link to optionally use the modern
attestation-backed APIs if available. Piped the
`link_mobile_use_attestation_endpoints` value from elements/sessions to
the Link components and API client. (But I'm unhappy that the signup
form has to know so much about the API calls!)
* Added a `deviceCanUseNativeLink` to tell us (synchronously) whether
the device is capable of attestation, and a `prepareNativeLink` to
attempt to attest a key and then return whether we can sign assertions.
We use these to help switch between the web and native flows.
* In the niche case where we're in the modern bindings and attestation
*doesn't* work, we'll disable signup completely.

### Errata
* Start sending the sessionID and Email source in Link lookup requests.
(These should have been sent before, and are required in the new
endpoints.)
* Stop logging out of Link automatically after an unsuccessful payment.
(This broke 3DS2 testing)
* Enabled attestation in the PaymentSheet Playground app's entitlements.

## Motivation
Enabling App Attestation in Link.

## Testing
Updated all impacted tests. Re-recorded network tests to include the new
`mobile_app_id` parameter.

## Changelog
Not yet released to users.