Attest separately for each API key and environment (#4378)
## Summary
We don't share attestation keys across API keys, so we need a separate
attestation key for each API key. This slightly increases the number of
attestations, but prevents situations where switching between merchants
causes attestations to fail until the hardcoded timeout.

We'll also relax the 24 hour limit when using the development App Attest
environment, as it doesn't increment the device's risk metric. We will
still enforce it for generating keys in the production environment. To
determine if we're in the development environment, we'll look for the
appattestdevelop aaguid in the attestation data blob.

## Motivation
Switching between merchants and environments without waiting for the 24
hour attestation failsafe to time out.

## Testing
Updated existing tests, added a new test, tested attestations on
physical device.

## Changelog
None