Fix issues with concurrent attestations (#4382)
## Summary
I realized that some of this code is unsafe if run concurrently, so I:
* Adopted actors to protect stored properties in StripeAttest and the
MockAppAttestService
* Used a stored `Task` to only allow one simultaneous attestation to
occur

Now we can kick off as many simultaneous `assert()` requests as we'd
like, and they all seem to complete without issues.

Note that I marked the appAttestService as `nonisolated` in order to
access the `isSupported` boolean synchronously. I _think_ this is fine
(DCAppAttestService appears to be set up to handle this correctly), but
this is our first use of actors in the SDK, so let me know if it seems
odd!

## Motivation
Fixing concurrency-related crashes and unexpected behavior

## Testing
Added a test, tested attestation on device

## Changelog
None