"Card details cannot be changed" text size closer to form label size (#4372)
## Summary
<!-- Simple summary of what was changed. -->
Changed smallFootnote to font size 11.5
The form label font is subheadline, but when the field has text or is
being edited, it moves up and scales to 75%. I tried making the
smallFootnote font subheadline and scaling it, but it scaled the label
position. 75% of 14pt is 10.5

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
https://jira.corp.stripe.com/browse/MOBILESDK-2878
## Testing
<!-- How was the code tested? Be as specific as possible. -->

![image](https://github.com/user-attachments/assets/65a6aad5-b1c2-4039-a3f1-84ff5c68d0cc)

![image](https://github.com/user-attachments/assets/7536533a-5850-4690-a5c0-42f57ac3feb7)

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->