Fix concurrent assertions (#4426)
## Summary
The server enforces that the [counter attached to the assertion
increases with each
assertion](https://developer.apple.com/documentation/devicecheck/validating-apps-that-connect-to-your-server?language=objc#Verify-the-assertion).
This means if one client happens to attempt to send multiple assertions
at once, and the network requests arrive out of order, we'll fail the
assertion.

To work around this, we'll change the API so that the `StripeAttest`
returns an `AssertionHandle`. It's required to call `complete()` with
this handle, or no future assertions will proceed.

I'm also going to make `stripeAttest` a parameter of an STPAPIClient
— now that there's more state involved, it makes more sense for an
STPAPIClient to own this instead of creating one each time.

## Testing
CI, added a new test, tested attestation on device.