Bail to web-based Link when needed (#4435)
## Summary
Add support for bailing back to the web Link flow when attestation
fails. This makes launching Link more responsive for the expected case
(that attestation will succeed if the app has attestation enabled and
the device supports attestation), while allowing a graceful fallback in
the unexpected case (the attestation or assertion fails). We'll bail if
either:
1. Initial attestation fails just after the sheet opens
2. If any asserted network request fails

## Motivation
Handle unexpected errors in a more user-friendly way.

## Testing
Tested manually by intentionally breaking assertion and attestation in
both cases.
Added a test that uses mocks to fail attestation and check that the
ASWebAuthenticationSession is presented.

## Changelog
None