Add interactive 2FA refresh for Link (#4432)
## Summary
We currently have some code to refresh the session using a token, but we
don't store a persistent login token in the app, so it won't work.
Instead, we'll pop the 2FA prompt again. This requires communicating
that state up from the PaymentSheetLinkAccount to the
PayWithLinkViewController so that we can present the authentication view
controller.

## Motivation
Allowing use of Link after the token expires.

## Testing
Added logic test, tested manually in PS Example.

## Changelog
None