Fix PaymentDetails CVC checks (#4456)
## Summary
`stateInvalid` is another state in which we should perform CVC
recollection.

## Testing
These states are tested `test_shouldRecollectCardCVC` test in
`PayWithLinkViewController_WalletViewModelTests`. Not sure if it's
useful to test all cases of this specific function, I've removed
`.default` so that the compiler will force us to reason about each case
in the future.