Fix linked libraries for FC Example app (#4474)
## Summary

This adds all the necessary libraries to the FC Example app. These are
all needed now that we added an MPE integration in our app, and fixes a
startup crash that was happening on physical devices.

## Motivation

Make the app usable

## Testing

Manual testing. Pushed a TestFlight build with these change too!

## Changelog

N/a