Break testAllIntegrationMethods into a few examples (#4473)
## Summary
This is pretty flaky, and most of these are tested in the PaymentSheet
LPM confirm tests. Let's just test a few to make sure the complex UI
stuff works (e.g. switching out to the Safari app then getting the
`UIApplication.didBecomeActiveNotification` when we switch back).

## Motivation
Reducing CI flake

## Testing
CI