Update promo in MPE after bank flow (#4370)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds code to refresh the MPE’s promo information after
the bank auth flow completes. If no incentive is available, the promo
badge in the carousel and bank form should be updated accordingly.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[CONSUMERBANK-571](https://jira.corp.stripe.com/browse/CONSUMERBANK-571)

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->