Refactor VerticalPaymentMethodListSelection -> RowButton.Type (#4526)
## Summary
This renames VerticalPaymentMethodListSelection to RowButton.Type and
adds a `type` property to RowButton.

## Motivation
There was potential value in having RowButton be agnostic to whether it
was a saved PM / new PM / Apple Pay / Link, in particular if we ever
wanted to move it to StripeCore. But it's too useful for RowButton to
hold onto its type and I don't think it's worse off for it. We can
simplify `EmbeddedPaymentMethodsView` a lot with this change.

Vertical mode could also benefit from this change but I'll leave that
for another PR.

## Testing
Existing tests, no new behavior

## Changelog
Not user facing