Fix a bug on setting presentingViewController in SwiftUI (#4556)
## Summary
- There was a bug where we would set the presentingViewController to one
of our own VCs which is incorrect and would cause the "Change" button to
no-op.
- The bug was caused in the following scenario: User fills out card,
hits continue, then hits change. When tapping change no form would be
presented, it would no-op.
- This was caused by `updateUIView` on our UIViewRepresentable being
called while our form VC is being dismissed. So when we compute the top
most VC, we get back our bottom sheet VC even though it's in the process
of dismissing.
- The fix is two fold: 1. Make sure we never set presenting VC to one of
our own. 2. If the top most VC is in the process of dismissing when
`updateUIView` is called, wait 0.1 seconds and try again.

## Motivation
- Embedded SwiftUI bug fix

## Testing
- Manual
- Updated UI test

## Changelog
N/A