Create subclasses of RowButton (#4594)
## Summary
- Makes all styles of RowButton a subclass and moves common properties
back into RowButton

Pros:
- Significantly DRYs up the style classes
- Easier to implement styles in the future
Cons:
- Accessibility is harder, we need a "contentView" to act as the
accessibility element for the RowButton since it cannot be it's own
accessibility element since we have a sub accessibility element
(accessory view). Content views are pretty normal so this feels fine.
- Introduces some weirdness w/ the RowButton init flagged below. Doesn't
seem too bad we have an assert preventing misuse.

Overall the DRY is nice and an improvement over the previous iteration.

## Motivation
- Reuse code for RowButton

## Testing
- Manual
- CI

## Changelog
N/A