Refactor CBC to use updateParams (#4589)
## Summary
* Refactors CBC so that it uses updateParams instead of using
`hasChangedCardBrand`.

* Removed `openCardBrandDropdown` and `closeCardBrandDropdown` as this
seems like too granular of analytics that I don't think we need --
instead it just logs cardBrandSelected, when the card brand changes.

Note: While removing implementing the cardBrandSelected in
`CardSectionElement`, i went down the path of changing all elements to
`lazy var xxx`, but felt this was too high risk of a change. This WAS
needed to debounce the number of metrics we were sending, but found a
way to do this with PaymentSheetAnalyticsHelper instead.

## Motivation
Refactoring code to make update payment method MM/YY, name on card,
eaiser.

## Testing
Relying on existing tests

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->