Make API calls to internal endpoint to set defaults (#4588)
## Summary
<!-- Simple summary of what was changed. -->
When updating a payment method, checking the checkbox calls the internal
endpoint to update the default payment method. Selecting a payment
method and pressing the action button in CustomerSheet calls the
internal endpoint to update the default payment method.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-2703](https://jira.corp.stripe.com/browse/MOBILESDK-2703)
[MOBILESDK-2701](https://jira.corp.stripe.com/browse/MOBILESDK-2701)
[MOBILESDK-2706](https://jira.corp.stripe.com/browse/MOBILESDK-2706)
## Testing
<!-- How was the code tested? Be as specific as possible. -->
testSetAsDefaultHorizontalNavigation_CustomerSession()
UI test with defaults and returning customer. No default set. Set card
ending in 4242 as the default payment method by checking the checkbox
and saving. Check that the card ending in 4242 is selected and has a
default badge. Pay, reload sheet, and check that the card ending in 4242
is still selected and has a default badge. Open the manage screen to
check that the checkbox is no longer displayed for the card ending in
4242 because it's already the default.

testSetAsDefaultVerticalNavigation_CustomerSession()
Same thing except with vertical

testAddTwoPaymentMethods_ChangeDefault()
CustomerSheet, add two cards. Select and confirm with the card that was
added first, check that it is now selected. Reload, see that it is still
selected.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->